inline some types
