--- src/api/response-types/image.ts
@@ -1,5 +1,0 @@
-export class ImageType {
-  id: string;
-  file: File;
-  sha256: string;
-}
