+++ src/api/response-types/role.ts
@@ -6,12 +6,3 @@ export class RoleType {
   permissions?: string[];
   locked: boolean;
 }
-
-export class GroupRoleType {
-  pulp_href: string;
-  pulp_created: string;
-  role: string;
-  description: string;
-  permissions?: string[];
-  content_object: string;
-}
