+++ src/containers/execution-environment-list/execution-environment-list.tsx
@@ -14,7 +14,6 @@ import { Link } from 'react-router-dom';
 import {
   ExecutionEnvironmentAPI,
   ExecutionEnvironmentRemoteAPI,
-  type ExecutionEnvironmentType,
 } from 'src/api';
 import { AppContext, type IAppContextType } from 'src/app-context';
 import {
@@ -50,6 +49,17 @@ import {
   withRouter,
 } from 'src/utilities';
 import './execution-environment.scss';
+
+interface ExecutionEnvironmentType {
+  created_at: string;
+  name: string;
+  description: string;
+  updated_at: string;
+  pulp: {
+    distribution: { base_path: string };
+    repository: { pulp_id: string; version: string };
+  };
+}
 
 interface IState {
   alerts: AlertType[];
